Added UserQuestions class
