+++ server/app/dao/QuestionsDao.java
@@ -4,7 +4,7 @@ import java.sql.*;
 import java.util.ArrayList;
 
 import play.db.*;
-import services.models.UserQuestion;
+import models.UserQuestion;
 
 /**
  * Created by sabinapokhrel on 11/7/16.
@@ -31,24 +31,26 @@ public class QuestionsDao {
 
   /**
    * This method contains the SQL statement to get stored data from questionsasked table.
-   *
+   * It returns an arrayList of UserQuestion objects.
    * @return
    */
-  public String getQuestions() {
+  public ArrayList<UserQuestion> getQuestions() {
 
     try (Connection connection = DB.getConnection()) {
       Statement statement = connection.createStatement();
       String sql = "SELECT * FROM questionsasked;";
       ResultSet resultSet = statement.executeQuery(sql);
 
-      ArrayList<String> questionList = new ArrayList<String>();
+      ArrayList<UserQuestion> questionList = new ArrayList<UserQuestion>();
       while (resultSet.next()) {
         String question = resultSet.getString("question");
-        questionList.add(question);
+        int questionNo = resultSet.getInt("questionNo");
+        UserQuestion userQuestion = new UserQuestion(questionNo, question);
+        questionList.add(userQuestion);
       }
       resultSet.close();
       statement.close();
-      return questionList.toString();
+      return questionList;
     } catch (Exception e) {
       e.printStackTrace();
       return null;
