--- server/app/services/models/BotResponse.java
@@ -1,11 +1,0 @@
-package services.models;
-
-/**
- * Created by sabinapokhrel on 10/1/16.
- *
- * BotResponse is the response sent to the user by the bot.
- */
-public class BotResponse {
-  public String status;
-  public String message;
-}
