--- server/app/services/models/UserQuestion.java
@@ -1,20 +1,0 @@
-package services.models;
-
-/**
- * Created by sabinapokhrel on 3/24/17.
- */
-public class UserQuestion {
-  public String getQuestion() {
-    return question;
-  }
-
-  public void setQuestion(String question) {
-    this.question = question;
-  }
-
-  private String question;
-
-  public UserQuestion(String question){
-    this.question = question;
-  }
-}
